requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.5.0
django-filter==24.3
djangorestframework==3.15.2
Markdown==3.7
pillow==10.4.0
sqlparse==0.5.1
tzdata==2024.1
